64bit compatibility paranoia

+++ common/FolderBrowser.h
@@ -2,8 +2,8 @@
 
 BOOL SkapaKatalog(char *szFile);
 int ValidatePath(char *szPath);
-int Browse(HWND hWnd,char *szFolder, int nFolder);
-int Browse(HWND hWnd,char *szFolder, int nFolder,char *szLabel,char *szValue,int nValue);
+intptr_t Browse(HWND hWnd,char *szFolder, int nFolder);
+intptr_t Browse(HWND hWnd,char *szFolder, int nFolder,char *szLabel,char *szValue,int nValue);
 int BrowseFile(HWND hwParent,BOOL bOpen, char *szTitle,char *szExt,char *szFile,int nFile);
 int BrowseFile(HWND hwParent,BOOL bOpen, char *szTitle,char *szExt,char *szFile,int nFile,char *szExtraName,int &nExtraValue,int nExtraMin, int nExtraMax);
 int BrowseFile(HWND hwParent,BOOL bOpen, char *szTitle,char *szExt,char *szFile,int nFile,char *szExtraCheck,int &nExtraCheck);
